# v7: more efficient NewSnapshot