--- a/v7/snapshot_test.go
+++ b/v7/snapshot_test.go
@@ -233,3 +233,29 @@ func TestLogging(t *testing.T) {
 		})
 	}
 }
+
+func TestNewSnapshot(t *testing.T) {
+	c := qt.New(t)
+	values := map[string]interface{}{
+		"intFlag":    1,
+		"floatFlag":  2.0,
+		"stringFlag": "three",
+		"boolFlag":   true,
+	}
+	snap, err := NewSnapshot(newTestLogger(t, LogLevelDebug), values)
+	c.Assert(err, qt.IsNil)
+	for key, want := range values {
+		c.Assert(snap.GetValue(key), qt.Equals, want)
+	}
+	// Sanity check that it works OK with Flag values.
+	c.Assert(Int("intFlag", 0).Get(snap), qt.Equals, 1)
+}
+
+func TestNewSnapshotWithUnknownType(t *testing.T) {
+	c := qt.New(t)
+	snap, err := NewSnapshot(newTestLogger(t, LogLevelDebug), map[string]interface{}{
+		"badVal": int64(1),
+	})
+	c.Check(err, qt.ErrorMatches, `value for flag "badVal" has unexpected type int64 \(1\); must be bool, int, float64 or string`)
+	c.Check(snap, qt.IsNil)
+}