### Describe the purpose of your pull request

Rather than creating a fake configuration just for the static snapshot,
we can just create the evaluators directly, which means that we can
do the creation faster and in O(1) allocations.

There's a small amount of performance regression in the other benchmarks,
but nothing really significant IMHO.

```
name                                   old time/op    new time/op    delta
Get/one-of/get-and-make-8                 272ns ± 0%     295ns ± 9%   +8.50%  (p=0.008 n=5+5)
Get/one-of/get-only-8                    53.4ns ± 0%    59.2ns ± 4%  +10.92%  (p=0.008 n=5+5)
Get/less-than-with-int/get-and-make-8     194ns ± 4%     209ns ± 1%   +7.24%  (p=0.008 n=5+5)
Get/less-than-with-int/get-only-8        22.6ns ± 0%    23.0ns ± 0%   +1.60%  (p=0.008 n=5+5)
Get/with-percentage/get-and-make-8        536ns ± 1%     555ns ± 1%   +3.46%  (p=0.008 n=5+5)
Get/with-percentage/get-only-8            322ns ± 0%     322ns ± 0%     ~     (p=0.643 n=5+5)
NewSnapshot-8                            41.3µs ± 2%     8.9µs ± 2%  -78.43%  (p=0.008 n=5+5)

name                                   old alloc/op   new alloc/op   delta
Get/one-of/get-and-make-8                  128B ± 0%      160B ± 0%  +25.00%  (p=0.008 n=5+5)
Get/one-of/get-only-8                     0.00B          0.00B          ~     (all equal)
Get/less-than-with-int/get-and-make-8     72.0B ± 0%    104.0B ± 0%  +44.44%  (p=0.008 n=5+5)
Get/less-than-with-int/get-only-8         0.00B          0.00B          ~     (all equal)
Get/with-percentage/get-and-make-8         200B ± 0%      232B ± 0%  +16.00%  (p=0.008 n=5+5)
Get/with-percentage/get-only-8            72.0B ± 0%     72.0B ± 0%     ~     (all equal)
NewSnapshot-8                            24.0kB ± 0%     4.6kB ± 0%  -80.67%  (p=0.008 n=5+5)

name                                   old allocs/op  new allocs/op  delta
Get/one-of/get-and-make-8                  2.00 ± 0%      2.00 ± 0%     ~     (all equal)
Get/one-of/get-only-8                      0.00           0.00          ~     (all equal)
Get/less-than-with-int/get-and-make-8      2.00 ± 0%      2.00 ± 0%     ~     (all equal)
Get/less-than-with-int/get-only-8          0.00           0.00          ~     (all equal)
Get/with-percentage/get-and-make-8         4.00 ± 0%      4.00 ± 0%     ~     (all equal)
Get/with-percentage/get-only-8             2.00 ± 0%      2.00 ± 0%     ~     (all equal)
NewSnapshot-8                               218 ± 0%         6 ± 0%  -97.25%  (p=0.008 n=5+5)
```

Also clarify the contract of `NewSnapshot` with repect to variation IDs and fix
it so that it doesn't return a slice of empty variation IDs.

### Requirement checklist (only if applicable)

- [x] I have covered the applied changes with automated tests.
- [x] I have executed the full automated test set against my changes.
- [ ] I have validated my changes against all supported platform versions.
- [x] I have read and accepted the [contribution agreement](https://github.com/configcat/legal/blob/main/contribution-agreement.md).